Preventing the nix build from using precompiled headers.
Some pre-compiled headers are used to speed up the nix build.

The nix build system is trying detect whether g++ or clang is in use
in order to point the compiler to the right pre-compiled headers. This
check obviously fails here since we are using afl-g++ as CXX compiler.

We don't really care about those pre-compiled headers in a fuzzing
context. Disabling their use altogether.